+++ scripts/update-types.sh
@@ -1,3 +1,3 @@
 #!/bin/sh
 # Should be executed with project root as the cwd
-npx tdl-install-types -o typings/tdlib-types.d.ts --flow prebuilt-tdlib
+exec npx tdl-install-types -o typings/tdlib-types.d.ts --flow prebuilt-tdlib